+++ src/Element/Html.php
@@ -34,27 +34,6 @@
 
 
     /**
-     * @deprecated
-     * @param $name
-     * @param $arguments
-     * @return $this|null
-     * @throws \Exception
-     */
-    public function __call($name, $arguments) {
-      trigger_error('Deprecated', E_USER_DEPRECATED);
-      if (strpos($name, 'set') === 0 and isset($arguments[0])) {
-        $name = strtolower(substr($name, 3));
-        $this->setAttribute($name, $arguments[0]);
-        return $this;
-      } elseif (strpos($name, 'get') === 0 and !isset($arguments[0])) {
-        return $this->getAttribute(strtolower(substr($name, 3)));
-      } else {
-        throw new \Exception('Invalid method: ' . $name);
-      }
-    }
-
-
-    /**
      * @param string $className
      * @return $this
      */